Case selection and particle communication bug
A new flag in the input file "Case=" allows to initialize the field data with the corresponding fields without having (forgeting) to compile the code.

In addition a known bug in the particle communication was corrected. The problem causes infinite loops when using non-periodic boundaries. The file corrected was Particles3Dcomm.cpp.

The Poisson corrector has been turned off because it causes problems in some configurations, slow down the code and does not give any tangible improvement to the results of the simulations.