New Poisson correction and Case flags + Particle depopulation
The new flag 'PoissonCorrection' allows to select if the Poisson correction must be active.

The flag 'Case' will select the initialization of the simulation. It is stored in Collective->Case and can be recovered using Collective::getCase(). It can be
 used in other places in the code for specific case related procedures. To add a new case modify the if/else if/else section in iPic3Dlib : c_Solver::Init().

Particle depopulator for the planet case has been included. It removes the particles inside a sphere and imposes a constant charge equal to rhoINIT.