Perfect conductor update + 1st version of the parallel HDF5
A new flag 'PoissonCorrection' allows to activate or desactivate the Poisson correction from the input file.

The prefectConductor boundary condition for the fields has been updated to cover all the 6 faces of the simulation domain.

The first version of the parallel HDF5 output was coded. New flags were added to the input file to use this output:
'WriteMethod' : if equal to 'Parallel' the new parallel output is used, with any other value the classical output method is used.
'SimulationName' : Gives a name to the simulation that is going to be used in the file name.

In the current commit only the Bxc field is saved. More fields will be added in the next commit.