Adding code for the coupling with BATSRUS
The files listed in this commit were modified to perform the coupling with the BATSRUS code.

The modifications are all encompased by a \#ifdef BATSRUS \#endif block.

To generate the makefiles for the BATSRUS version (activation of the -DBATSRUS flag) in your build directory type:
BATSRUS=1 cmake <src/directory>
where <src/directory> is the directory containing the CMakeLists.txt file.

In addition a new case 'BATSRUS' was included for this version. When running the code the 'Case' flag in the inputfile must be equal to 'BATSRUS' to activate the correct field and particle initializaion (see main/iPic3Dlib.cpp and the new initBATSRUS in fields/EMfields3D.cpp).