Deactivation of Parallel HDF5 options
By default the code now compiles using the sequential version of the HDF5 library. To use the parallel version make sure that the Parallel HDF5 version is available in the system and use the following command when configuring the code in the installation directory:

USE_PHDF5=1 cmake <src/dir>

where <src/dir> is the path to the source directory of the code.

NOTE: The same method is used to activate the BATSRUS portions of the code. So if you're planning to use the Parallel HDF5 and BATSRUS you must type:

USE_PHDF5=1 BATSRUS=1 cmake <src/dir>