Optional compilation of library as shared
In some systems, depending on the installed libraries configuration, a shared library can not be compiled. In such case the compilation must be done using static libraries. STATIC is now given by default, but shared libraries can be activated with the cmake option -DIPIC_BUILD_SHARED=ON or using the ccmake GUI.