Bug correction: wrong BC treatment
The new test tool showed that a bug was inserted at some point producing an incorrect run of the GEM test.

The error was traced back to commit 01e8f3c15f38f161b90915bec6e77bf1c24bc25f where a constant charge was applied not only to OpenBC conditions, but also to perfect conductor conditions, which caused a bad behavior of the code in the GEM run.