Moving calculateB to the field subroutine
The call to calculateB was still performed after the particle mover inside the particle subroutine. Moving this call causes the particle mover to use the updated (n+1) magnetic field, instead of the previous (n) B field. The tests show a very small difference between the two versions which increasses over time.

The test script has been updated to lower the failure threshold. The minimum relative error was set to 0.1% and is only calculated if the absolute error is above 1e-7.