Parallel HDF5 is now done using the H5hut library
Reading and writing functions that use the H5hut library have been included. These functions can be used in iPic3D but are also useful in pre- and post-processing tools. For such reason the H5hut library is handled by the new library saved in the H5hut-io directory.

The CMake structure of the code has been changed: the 'main' CMakeLists.txt file is located in the root directory. This file calls cmake inside the main directory and inside the H5hut-io directory if it is required by the user. It also performs the necessary links between all the libraries.

MPI functions were included to recover information about the topology.
Functions to gather information at the cell center have been included in the fields side of the code.