IMPORTANT: bug correction
From issue #56: the iterative particle mover was doing only 1 iteration per cycle.

Automatic testing reflected inaccuracies between the reference simulation and the current version. The order in which the main code calls were made was incorrect. In this commit we re-wrote the algorithm in the correct order.

The -O3 flag was added by default to the Makefile so the code can run at top speed.