Bug correction
When calculating the maximum number of particles, the code was using the total number of cells in the domain times the number of particles per cell, and then the total was divided by the number of processes.

In very big cases this leads to extremely large numbers, larger than the limit of the 'long' and 'long long' types. The code was moified to calculate the maximum number of particles per process, to avoid such numbers.