Injected particles in random positions
Injecting particles in the inflow boundaries must be done using a random placement of the particles. If it is done in homogeneous locations the shared nodes between the injecting cells and the domain cells get double/half the density that is expected. Only using a random location of particles the inlet works correctly.

However, we still have an uneven charge density and that can also lead to an uneven current (particle velocities). It is important to use as many particles per cell as possible, at least 8x8x8. This problem must be studied in the future.