Adding particle initialization using thermal velocities from file
Reading and writing the three components of the thermal velocities Vthx, Vthy, and Vthz.
This fields can be used to restart a simulation. An additional particle routine is included to perform such restart.
Options have been added to the input file to perform the restart and to request the code to save the thermal velocity fields.